# postman/postman-collection-openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/products": {
            "post": {
                "description": "Add a new product to the catalog. Requires admin or moderator role.",
                "operationId": "create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateProductDto" } } }
                },
                "responses": {
                    "201": {
                        "description": "Product successfully created",
                        "content": {
                            "application/json": {
                                "example": {
                                    "product_id": 1,
                                    "name": "iPhone 15 Pro",
                                    "description": "Latest iPhone with advanced features",
                                    "price": 1299.99,
                                    "category": "Electronics",
                                    "brand": "Apple",
                                    "stock_quantity": 10,
                                    "image_url": "https://example.com/iphone15pro.jpg",
                                    "is_active": true,
                                    "created_at": "2025-07-03T03:14:08.000Z",
                                    "updated_at": "2025-07-03T03:14:08.000Z"
                                }
                            }
                        }
                    },
                    "400": { "description": "Invalid input data" },
                    "401": { "description": "Unauthorized - Invalid or missing token" },
                    "403": { "description": "Forbidden - Insufficient permissions" }
                },
                "security": [{ "bearer": [] }],
                "summary": "Create a new product (Admin/Moderator only)",
                "tags": ["Products"]
            },
            "get": {
                "description": "Retrieve products with pagination, filtering, sorting, searching, and includes. Supports price ranges, category filtering, brand filtering, stock level filtering, and more.",
                "operationId": "findAll",
                "parameters": [
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "description": "Page number for pagination",
                        "schema": { "minimum": 1, "default": 1, "example": 1, "type": "number" }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "description": "Number of results per page",
                        "schema": { "minimum": 1, "maximum": 100, "default": 10, "example": 10, "type": "number" }
                    },
                    {
                        "name": "offset",
                        "required": false,
                        "in": "query",
                        "description": "Number of records to skip (overrides page calculation)",
                        "schema": { "minimum": 0, "example": 0, "type": "number" }
                    },
                    {
                        "name": "sortBy",
                        "required": false,
                        "in": "query",
                        "description": "Field to sort by",
                        "schema": { "example": "created_at", "type": "string" }
                    },
                    {
                        "name": "sortOrder",
                        "required": false,
                        "in": "query",
                        "description": "Sort order",
                        "schema": { "default": "ASC", "example": "ASC", "type": "string", "enum": ["ASC", "DESC"] }
                    },
                    {
                        "name": "search",
                        "required": false,
                        "in": "query",
                        "description": "Search term to filter results",
                        "schema": { "example": "laptop", "type": "string" }
                    },
                    {
                        "name": "include",
                        "required": false,
                        "in": "query",
                        "description": "Related entities to include in the response (joins). Can be comma-separated string or array",
                        "schema": { "example": ["user", "product"], "type": "array", "items": { "type": "array" } }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Products retrieved successfully",
                        "content": {
                            "application/json": {
                                "example": {
                                    "data": [
                                        {
                                            "product_id": 1,
                                            "name": "iPhone 15 Pro",
                                            "description": "Latest iPhone with advanced features",
                                            "price": 1299.99,
                                            "category": "Electronics",
                                            "brand": "Apple",
                                            "stock_quantity": 10,
                                            "image_url": "https://example.com/iphone15pro.jpg",
                                            "is_active": true,
                                            "created_at": "2025-07-03T03:14:08.000Z",
                                            "updated_at": "2025-07-03T03:14:08.000Z"
                                        }
                                    ],
                                    "pagination": { "page": 1, "limit": 10, "total": 1, "totalPages": 1, "hasNext": false, "hasPrev": false }
                                }
                            }
                        }
                    }
                },
                "summary": "Get all products with advanced querying",
                "tags": ["Products"]
            }
        },
        "/products/category/{category}": {
            "get": {
                "operationId": "findByCategory",
                "parameters": [{ "name": "category", "required": true, "in": "path", "schema": { "type": "string" } }],
                "responses": { "200": { "description": "" } },
                "tags": ["Products"]
            }
        },
        "/products/{id}": {
            "get": {
                "operationId": "findOne",
                "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
                "responses": { "200": { "description": "" } },
                "tags": ["Products"]
            },
            "patch": {
                "description": "Update product information. Requires admin or moderator role.",
                "operationId": "update",
                "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
                "requestBody": {
                    "required": true,
                    "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateProductDto" } } }
                },
                "responses": {
                    "401": { "description": "Unauthorized - Invalid or missing token" },
                    "403": { "description": "Forbidden - Insufficient permissions" }
                },
                "security": [{ "bearer": [] }],
                "summary": "Update product (Admin/Moderator only)",
                "tags": ["Products"]
            },
            "delete": {
                "description": "Delete a product from the catalog. Only admins can delete products.",
                "operationId": "remove",
                "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
                "responses": {
                    "401": { "description": "Unauthorized - Invalid or missing token" },
                    "403": { "description": "Forbidden - Admin access required" }
                },
                "security": [{ "bearer": [] }],
                "summary": "Delete product (Admin only)",
                "tags": ["Products"]
            }
        },
        "/products/{id}/stock": {
            "patch": {
                "description": "Update product stock quantity. Requires admin or moderator role.",
                "operationId": "updateStock",
                "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
                "responses": {
                    "401": { "description": "Unauthorized - Invalid or missing token" },
                    "403": { "description": "Forbidden - Insufficient permissions" }
                },
                "security": [{ "bearer": [] }],
                "summary": "Update product stock (Admin/Moderator only)",
                "tags": ["Products"]
            }
        },
...
        "/reviews/product/{productId}": {
            "get": {
                "description": "Retrieve all reviews for a given product ID",
                "operationId": "findByProduct",
                "parameters": [
                    { "name": "productId", "required": true, "in": "path", "description": "Product ID", "schema": { "example": 1, "type": "number" } }
                ],
                "responses": {
                    "200": {
                        "description": "Product reviews retrieved successfully",
                        "content": {
                            "application/json": {
                                "example": [
                                    {
                                        "id": 1,
                                        "user_id": 1,
                                        "product_id": 1,
                                        "rating": 5,
                                        "title": "Great product!",
                                        "comment": "This product exceeded my expectations.",
                                        "is_verified_purchase": true,
                                        "helpful_votes": 5,
                                        "created_at": "2024-01-01T00:00:00Z",
                                        "updated_at": "2024-01-01T00:00:00Z"
                                    }
                                ]
                            }
                        }
                    }
                },
                "summary": "Get reviews for a specific product",
                "tags": ["Reviews"]
            }
        },
...
    },
    "info": {
        "title": "eCommerce API",
        "description": "Comprehensive eCommerce API with advanced querying capabilities including pagination, filtering, sorting, searching, and joins",
        "version": "1.0.0",
        "contact": { "name": "API Support", "url": "https://github.com/yourusername/ecommerce-api", "email": "support@yourcompany.com" },
        "license": { "name": "MIT", "url": "https://opensource.org/licenses/MIT" }
    },
...
    "servers": [{ "url": "http://localhost:3151", "description": "Development server" }],
    "components": {
        "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
        "schemas": {
...
            "CreateProductDto": {
                "type": "object",
                "properties": {
                    "name": { "type": "string", "example": "iPhone 15 Pro", "description": "Product name", "maxLength": 255 },
                    "description": {
                        "type": "string",
                        "example": "Latest iPhone with A17 Pro chip, titanium design, and pro camera system",
                        "description": "Detailed product description"
                    },
                    "price": { "type": "number", "example": 1299.99, "description": "Product price in CAD", "minimum": 0 },
                    "category": { "type": "string", "example": "Electronics", "description": "Product category", "maxLength": 100 },
                    "brand": { "type": "string", "example": "Apple", "description": "Product brand", "maxLength": 100 },
                    "stock_quantity": { "type": "number", "example": 50, "description": "Available stock quantity", "minimum": 0 },
                    "image_url": {
                        "type": "string",
                        "example": "https://example.com/images/iphone15pro.jpg",
                        "description": "Product image URL",
                        "maxLength": 500
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true,
                        "description": "Whether the product is active/available for sale",
                        "default": true
                    }
                },
                "required": ["name", "price", "category"]
            },
            "UpdateProductDto": {
                "type": "object",
                "properties": {
                    "name": { "type": "string", "example": "iPhone 15 Pro Max", "description": "Product name", "maxLength": 255 },
                    "description": {
                        "type": "string",
                        "example": "Latest iPhone with A17 Pro chip, titanium design, and pro camera system with 5x zoom",
                        "description": "Detailed product description"
                    },
                    "price": { "type": "number", "example": 1499.99, "description": "Product price in CAD", "minimum": 0 },
                    "category": { "type": "string", "example": "Electronics", "description": "Product category", "maxLength": 100 },
                    "brand": { "type": "string", "example": "Apple", "description": "Product brand", "maxLength": 100 },
                    "stock_quantity": { "type": "number", "example": 25, "description": "Available stock quantity", "minimum": 0 },
                    "image_url": {
                        "type": "string",
                        "example": "https://example.com/images/iphone15promax.jpg",
                        "description": "Product image URL",
                        "maxLength": 500
                    },
                    "is_active": { "type": "boolean", "example": true, "description": "Whether the product is active/available for sale" }
                }
            },
...
        }
    }
}